+++ corehq/apps/app_manager/detail_screen.py
@@ -1,5 +1,6 @@
 from corehq.apps.app_manager import id_strings
-from corehq.apps.app_manager import suite_xml as sx
+from corehq.apps.app_manager.suite_xml import models as sx
+from corehq.apps.app_manager.suite_xml import const
 from corehq.apps.app_manager.util import is_sort_only_column
 from corehq.apps.app_manager.xpath import (
     CaseXPath,
@@ -466,14 +467,14 @@ class Graph(FormattedDetailColumn):
         return template
 
 
-@register_type_processor(sx.FIELD_TYPE_ATTACHMENT)
+@register_type_processor(const.FIELD_TYPE_ATTACHMENT)
 class AttachmentXpathGenerator(BaseXpathGenerator):
     @property
     def xpath(self):
-        return sx.FIELD_TYPE_ATTACHMENT + "/" + self.column.field_property
+        return const.FIELD_TYPE_ATTACHMENT + "/" + self.column.field_property
 
 
-@register_type_processor(sx.FIELD_TYPE_PROPERTY)
+@register_type_processor(const.FIELD_TYPE_PROPERTY)
 class PropertyXpathGenerator(BaseXpathGenerator):
     @property
     def xpath(self):
@@ -515,7 +516,7 @@ class PropertyXpathGenerator(BaseXpathGenerator):
         )
 
 
-@register_type_processor(sx.FIELD_TYPE_INDICATOR)
+@register_type_processor(const.FIELD_TYPE_INDICATOR)
 class IndicatorXpathGenerator(BaseXpathGenerator):
     @property
     def xpath(self):
@@ -524,7 +525,7 @@ class IndicatorXpathGenerator(BaseXpathGenerator):
         return IndicatorXpath(instance_id).instance().slash(indicator)
 
 
-@register_type_processor(sx.FIELD_TYPE_LOCATION)
+@register_type_processor(const.FIELD_TYPE_LOCATION)
 class LocationXpathGenerator(BaseXpathGenerator):
     @property
     def xpath(self):
@@ -533,7 +534,7 @@ class LocationXpathGenerator(BaseXpathGenerator):
         return LocationXpath('commtrack:locations').location(self.column.field_property, hierarchy)
 
 
-@register_type_processor(sx.FIELD_TYPE_LEDGER)
+@register_type_processor(const.FIELD_TYPE_LEDGER)
 class LedgerXpathGenerator(BaseXpathGenerator):
 
     @property
@@ -549,7 +550,7 @@ class LedgerXpathGenerator(BaseXpathGenerator):
         )
 
 
-@register_type_processor(sx.FIELD_TYPE_SCHEDULE)
+@register_type_processor(const.FIELD_TYPE_SCHEDULE)
 class ScheduleXpathGenerator(BaseXpathGenerator):
 
     @property
