+++ corehq/apps/app_manager/models.py
@@ -25,6 +25,7 @@ from django.utils.encoding import force_unicode
 from django.utils.safestring import mark_safe
 from django.utils.translation import override, ugettext as _, ugettext
 from couchdbkit.exceptions import BadValueError
+from corehq.apps.app_manager.suite_xml_old import SuiteGenerator, MediaSuiteGenerator
 from corehq.util.soft_assert import soft_assert
 from dimagi.ext.couchdbkit import *
 from django.conf import settings
@@ -64,7 +65,8 @@ from corehq.apps.users.models import CouchUser
 from corehq.apps.users.util import cc_user_domain
 from corehq.apps.domain.models import cached_property, Domain
 from corehq.apps.app_manager import current_builds, app_strings, remote_app, \
-    id_strings, suite_xml, commcare_settings
+    id_strings, commcare_settings
+from corehq.apps.app_manager.suite_xml import models as suite_models
 from corehq.apps.app_manager.dbaccessors import get_app
 from corehq.apps.app_manager.util import (
     split_path,
@@ -1844,7 +1846,7 @@ class ModuleBase(IndexedSchema, NavMenuItemMediaMixin):
             detail.rename_lang(old_lang, new_lang)
 
     def validate_detail_columns(self, columns):
-        from corehq.apps.app_manager.suite_xml import FIELD_TYPE_LOCATION
+        from corehq.apps.app_manager.suite_xml.const import FIELD_TYPE_LOCATION
         from corehq.apps.locations.util import parent_child
         hierarchy = None
         for column in columns:
@@ -2655,7 +2657,7 @@ class AdvancedModule(ModuleBase):
 
                 if from_module.parent_select.active:
                     app = self.get_app()
-                    select_chain = suite_xml.SuiteGenerator.get_select_chain(app, from_module, include_self=False)
+                    select_chain = SuiteGenerator.get_select_chain(app, from_module, include_self=False)
                     for n, link in enumerate(reversed(list(enumerate(select_chain)))):
                         i, module = link
                         new_form.actions.load_update_cases.append(LoadUpdateAction(
@@ -3332,21 +3334,21 @@ class ReportAppConfig(DocumentSchema):
         yield (self.summary_detail_id, self.summary_details(), True)
 
     def select_details(self):
-        return Detail(custom_xml=suite_xml.Detail(
+        return Detail(custom_xml=suite_models.Detail(
             id='reports.{}.select'.format(self.uuid),
-            title=suite_xml.Text(
-                locale=suite_xml.Locale(id=id_strings.report_menu()),
+            title=suite_models.Text(
+                locale=suite_models.Locale(id=id_strings.report_menu()),
             ),
             fields=[
-                suite_xml.Field(
-                    header=suite_xml.Header(
-                        text=suite_xml.Text(
-                            locale=suite_xml.Locale(id=id_strings.report_name_header()),
+                suite_models.Field(
+                    header=suite_models.Header(
+                        text=suite_models.Text(
+                            locale=suite_models.Locale(id=id_strings.report_name_header()),
                         )
                     ),
-                    template=suite_xml.Template(
-                        text=suite_xml.Text(
-                            locale=suite_xml.Locale(id=id_strings.report_name(self.uuid))
+                    template=suite_models.Template(
+                        text=suite_models.Text(
+                            locale=suite_models.Locale(id=id_strings.report_name(self.uuid))
                         )
                     ),
                 )
@@ -3362,75 +3364,75 @@ class ReportAppConfig(DocumentSchema):
                     graph_config = self.graph_configs.get(chart_config.chart_id, ReportGraphConfig())
 
                     def _column_to_series(column):
-                        return suite_xml.Series(
+                        return suite_models.Series(
                             nodeset="instance('reports')/reports/report[@id='{}']/rows/row".format(self.uuid),
                             x_function="column[@id='{}']".format(chart_config.x_axis_column),
                             y_function="column[@id='{}']".format(column),
-                            configuration=suite_xml.ConfigurationGroup(configs=[
-                                suite_xml.ConfigurationItem(id=key, xpath_function=value)
+                            configuration=suite_models.ConfigurationGroup(configs=[
+                                suite_models.ConfigurationItem(id=key, xpath_function=value)
                                 for key, value in graph_config.series_configs.get(column, {}).items()
                             ])
                         )
-                    yield suite_xml.Field(
-                        header=suite_xml.Header(text=suite_xml.Text()),
-                        template=suite_xml.GraphTemplate(
+                    yield suite_models.Field(
+                        header=suite_models.Header(text=suite_models.Text()),
+                        template=suite_models.GraphTemplate(
                             form='graph',
-                            graph=suite_xml.Graph(
+                            graph=suite_models.Graph(
                                 type=graph_config.graph_type,
                                 series=[_column_to_series(c) for c in chart_config.y_axis_columns],
-                                configuration=suite_xml.ConfigurationGroup(configs=[
-                                    suite_xml.ConfigurationItem(id=key, xpath_function=value)
+                                configuration=suite_models.ConfigurationGroup(configs=[
+                                    suite_models.ConfigurationItem(id=key, xpath_function=value)
                                     for key, value in graph_config.config.items()
                                 ]),
                             ),
                         )
                     )
 
-        return Detail(custom_xml=suite_xml.Detail(
+        return Detail(custom_xml=suite_models.Detail(
             id='reports.{}.summary'.format(self.uuid),
-            title=suite_xml.Text(
-                locale=suite_xml.Locale(id=id_strings.report_menu()),
+            title=suite_models.Text(
+                locale=suite_models.Locale(id=id_strings.report_menu()),
             ),
             details=[
-                suite_xml.Detail(
-                    title=suite_xml.Text(
-                        locale=suite_xml.Locale(id=id_strings.report_menu()),
+                suite_models.Detail(
+                    title=suite_models.Text(
+                        locale=suite_models.Locale(id=id_strings.report_menu()),
                     ),
                     fields=[
-                        suite_xml.Field(
-                            header=suite_xml.Header(
-                                text=suite_xml.Text(
-                                    locale=suite_xml.Locale(id=id_strings.report_name_header())
+                        suite_models.Field(
+                            header=suite_models.Header(
+                                text=suite_models.Text(
+                                    locale=suite_models.Locale(id=id_strings.report_name_header())
                                 )
                             ),
-                            template=suite_xml.Template(
-                                text=suite_xml.Text(
-                                    locale=suite_xml.Locale(id=id_strings.report_name(self.uuid))
+                            template=suite_models.Template(
+                                text=suite_models.Text(
+                                    locale=suite_models.Locale(id=id_strings.report_name(self.uuid))
                                 )
                             ),
                         ),
-                        suite_xml.Field(
-                            header=suite_xml.Header(
-                                text=suite_xml.Text(
-                                    locale=suite_xml.Locale(id=id_strings.report_description_header()),
+                        suite_models.Field(
+                            header=suite_models.Header(
+                                text=suite_models.Text(
+                                    locale=suite_models.Locale(id=id_strings.report_description_header()),
                                 )
                             ),
-                            template=suite_xml.Template(
-                                text=suite_xml.Text(
-                                    xpath=suite_xml.Xpath(function='description')
+                            template=suite_models.Template(
+                                text=suite_models.Text(
+                                    xpath=suite_models.Xpath(function='description')
                                 )
                             ),
                         ),
                     ] + list(_get_graph_fields()) + [
-                        suite_xml.Field(
-                            header=suite_xml.Header(
-                                text=suite_xml.Text(
-                                    locale=suite_xml.Locale(id=id_strings.report_last_sync())
+                        suite_models.Field(
+                            header=suite_models.Header(
+                                text=suite_models.Text(
+                                    locale=suite_models.Locale(id=id_strings.report_last_sync())
                                 )
                             ),
-                            template=suite_xml.Template(
-                                text=suite_xml.Text(
-                                    xpath=suite_xml.Xpath(
+                            template=suite_models.Template(
+                                text=suite_models.Text(
+                                    xpath=suite_models.Xpath(
                                         function="format-date(date(instance('reports')/reports/@last_sync), '%Y-%m-%d %H:%M')"
                                     )
                                 )
@@ -3444,39 +3446,39 @@ class ReportAppConfig(DocumentSchema):
 
     def data_detail(self):
         def _column_to_field(column):
-            return suite_xml.Field(
-                header=suite_xml.Header(
-                    text=suite_xml.Text(
-                        locale=suite_xml.Locale(
+            return suite_models.Field(
+                header=suite_models.Header(
+                    text=suite_models.Text(
+                        locale=suite_models.Locale(
                             id=id_strings.report_column_header(self.uuid, column.column_id)
                         ),
                     )
                 ),
-                template=suite_xml.Template(
-                    text=suite_xml.Text(
-                        xpath=suite_xml.Xpath(function="column[@id='{}']".format(column.column_id)))
+                template=suite_models.Template(
+                    text=suite_models.Text(
+                        xpath=suite_models.Xpath(function="column[@id='{}']".format(column.column_id)))
                 ),
             )
 
-        return suite_xml.Detail(
+        return suite_models.Detail(
             id='reports.{}.data'.format(self.uuid),
             nodeset='rows/row',
-            title=suite_xml.Text(
-                locale=suite_xml.Locale(id=id_strings.report_data_table()),
+            title=suite_models.Text(
+                locale=suite_models.Locale(id=id_strings.report_data_table()),
             ),
             fields=[_column_to_field(c) for c in self.report.report_columns]
         )
 
     def get_entry(self):
-        return suite_xml.Entry(
-            command=suite_xml.Command(
+        return suite_models.Entry(
+            command=suite_models.Command(
                 id='reports.{}'.format(self.uuid),
-                text=suite_xml.Text(
-                    locale=suite_xml.Locale(id=id_strings.report_name(self.uuid)),
+                text=suite_models.Text(
+                    locale=suite_models.Locale(id=id_strings.report_name(self.uuid)),
                 ),
             ),
             datums=[
-                suite_xml.SessionDatum(
+                suite_models.SessionDatum(
                     detail_confirm=self.summary_detail_id,
                     detail_select=self.select_detail_id,
                     id='report_id_{}'.format(self.uuid),
@@ -3535,13 +3537,13 @@ class ReportModule(ModuleBase):
             yield config.get_entry()
 
     def get_menus(self):
-        yield suite_xml.Menu(
+        yield suite_models.Menu(
             id=id_strings.menu_id(self),
-            text=suite_xml.Text(
-                locale=suite_xml.Locale(id=id_strings.module_locale(self))
+            text=suite_models.Text(
+                locale=suite_models.Locale(id=id_strings.module_locale(self))
             ),
             commands=[
-                suite_xml.Command(id=id_strings.report_command(config.uuid))
+                suite_models.Command(id=id_strings.report_command(config.uuid))
                 for config in self.report_configs
             ]
         )
@@ -4571,10 +4573,10 @@ class Application(ApplicationBase, TranslationMixin, HQMediaMixin):
                 'langs': ["default"] + self.build_langs
             })
         else:
-            return suite_xml.SuiteGenerator(self).generate_suite()
+            return SuiteGenerator(self).generate_suite()
 
     def create_media_suite(self):
-        return suite_xml.MediaSuiteGenerator(self).generate_suite()
+        return MediaSuiteGenerator(self).generate_suite()
 
     @classmethod
     def get_form_filename(cls, type=None, form=None, module=None):
