+++ corehq/apps/app_manager/suite_xml_old.py
@@ -9,7 +9,7 @@ from xml.sax.saxutils import escape
 from django.utils.translation import ugettext_noop as _
 from django.core.urlresolvers import reverse
 from corehq.apps.app_manager.suite_xml.const import FIELD_TYPE_SCHEDULE
-from corehq.apps.app_manager.suite_xml.details import get_detail_column_infos
+from corehq.apps.app_manager.suite_xml.details import get_detail_column_infos, DetailContributor
 from corehq.apps.app_manager.suite_xml.workflow import WorkflowHelper
 
 from .exceptions import (
@@ -178,78 +178,6 @@ class SuiteGenerator(SuiteGeneratorBase):
                 resource.descriptor = u"Translations: %s" % languages_mapping().get(lang, [unknown_lang_txt])[0]
             yield resource
 
-    def build_detail(self, module, detail_type, detail, detail_column_infos,
-                     tabs, id, title, start, end):
-        """
-        Recursively builds the Detail object.
-        (Details can contain other details for each of their tabs)
-        """
-        from corehq.apps.app_manager.detail_screen import get_column_generator
-        d = Detail(id=id, title=title)
-        if tabs:
-            tab_spans = detail.get_tab_spans()
-            for tab in tabs:
-                sub_detail = self.build_detail(
-                    module,
-                    detail_type,
-                    detail,
-                    detail_column_infos,
-                    [],
-                    None,
-                    Text(locale_id=id_strings.detail_tab_title_locale(
-                        module, detail_type, tab
-                    )),
-                    tab_spans[tab.id][0],
-                    tab_spans[tab.id][1]
-                )
-                if sub_detail:
-                    d.details.append(sub_detail)
-            if len(d.details):
-                return d
-            else:
-                return None
-
-        # Base case (has no tabs)
-        else:
-            # Add lookup
-            if detail.lookup_enabled and detail.lookup_action:
-                d.lookup = Lookup(
-                    name=detail.lookup_name or None,
-                    action=detail.lookup_action,
-                    image=detail.lookup_image or None,
-                )
-                d.lookup.extras = [Extra(**e) for e in detail.lookup_extras]
-                d.lookup.responses = [Response(**r) for r in detail.lookup_responses]
-
-            # Add variables
-            variables = list(
-                SuiteGenerator.detail_variables(module, detail, detail_column_infos[start:end])
-            )
-            if variables:
-                d.variables.extend(variables)
-
-            # Add fields
-            for column_info in detail_column_infos[start:end]:
-                fields = get_column_generator(
-                    self.app, module, detail,
-                    detail_type=detail_type, *column_info
-                ).fields
-                d.fields.extend(fields)
-
-            # Add actions
-            if module.case_list_form.form_id and detail_type.endswith('short')\
-                    and not module.put_in_root:
-                self._add_action_to_detail(d, module)
-
-            try:
-                if not self.app.enable_multi_sort:
-                    d.fields[0].sort = 'default'
-            except IndexError:
-                pass
-            else:
-                # only yield the Detail if it has Fields
-                return d
-
     def get_datums_meta_for_form_generic(self, form):
         if form.form_type == 'module_form':
             datums_meta = self.get_case_datums_basic_module(form.get_module(), form)
@@ -318,180 +246,7 @@ class SuiteGenerator(SuiteGeneratorBase):
     @property
     @memoized
     def details(self):
-        r = []
-        if not self.app.use_custom_suite:
-            for module in self.modules:
-                for detail_type, detail, enabled in module.get_details():
-                    if enabled:
-                        if detail.custom_xml:
-                            d = load_xmlobject_from_string(
-                                detail.custom_xml,
-                                xmlclass=Detail
-                            )
-                            r.append(d)
-                        else:
-                            detail_column_infos = get_detail_column_infos(
-                                detail,
-                                include_sort=detail_type.endswith('short'),
-                            )
-                            if detail_column_infos:
-                                if detail.use_case_tiles:
-                                    r.append(self.build_case_tile_detail(
-                                        module, detail, detail_type
-                                    ))
-                                else:
-                                    d = self.build_detail(
-                                        module,
-                                        detail_type,
-                                        detail,
-                                        detail_column_infos,
-                                        list(detail.get_tabs()),
-                                        id_strings.detail(module, detail_type),
-                                        Text(locale_id=id_strings.detail_title_locale(
-                                            module, detail_type
-                                        )),
-                                        0,
-                                        len(detail_column_infos)
-                                    )
-                                    if d:
-                                        r.append(d)
-                if module.fixture_select.active:
-                    d = Detail(
-                        id=id_strings.fixture_detail(module),
-                        title=Text(),
-                    )
-                    xpath = Xpath(function=module.fixture_select.display_column)
-                    if module.fixture_select.localize:
-                        template_text = Text(locale=Locale(child_id=Id(xpath=xpath)))
-                    else:
-                        template_text = Text(xpath_function=module.fixture_select.display_column)
-                    fields = [Field(header=Header(text=Text()),
-                                    template=Template(text=template_text),
-                                    sort_node='')]
-
-                    d.fields = fields
-                    r.append(d)
-        return r
-
-    @staticmethod
-    def _schedule_detail_variables(module, detail, detail_column_infos):
-        has_schedule_columns = any(ci.column.field_type == FIELD_TYPE_SCHEDULE for ci in detail_column_infos)
-        has_schedule = getattr(module, 'has_schedule', False)
-        if (has_schedule and module.all_forms_require_a_case() and has_schedule_columns):
-            yield DetailVariable(name=SCHEDULE_DATE_CASE_OPENED, function="date_opened")  # date case is opened
-            forms_due = []
-            last_visit_dates = []
-            for phase in module.get_schedule_phases():
-                if not phase.anchor:
-                    raise ScheduleError(_("Schedule Phase in module '{module_name}' is missing an anchor")
-                                        .format(module_name=module.default_name()))
-
-                for form in phase.get_forms():
-                    """
-                    Adds the following variables for each form:
-                    <anchor_{form_id} function="{anchor}"/>
-                    <last_visit_number_{form_id} function="{last_visit_number}"/>
-                    <last_visit_date_{form_id} function="{last_visit_date}"/>
-                    <next_{form_id} function={phase_set}/>
-                    """
-                    if not form.schedule_form_id:
-                        raise ScheduleError(
-                            _("Form '{form_name}' in module '{module_name}' is missing an abbreviation")
-                            .format(form_name=trans(form["name"], langs=[module.get_app().default_language]),
-                                    module_name=module.default_name()))
-                    form_xpath = ScheduleFormXPath(form, phase, module)
-                    name = u"next_{}".format(form.schedule_form_id)
-                    forms_due.append(u"${}".format(name))
-
-                    # Add an anchor and last_visit variables so we can reference it in the calculation
-                    yield DetailVariable(name=form_xpath.anchor_detail_variable_name, function=phase.anchor)
-                    yield DetailVariable(name=form_xpath.last_visit_detail_variable_name,
-                                         function=SCHEDULE_LAST_VISIT.format(form.schedule_form_id))
-                    yield DetailVariable(name=form_xpath.last_visit_date_detail_variable_name,
-                                         function=SCHEDULE_LAST_VISIT_DATE.format(form.schedule_form_id))
-                    last_visit_dates.append(form_xpath.last_visit_date_detail_variable_name)
-                    if phase.id == 1:
-                        # If this is the first phase, `current_schedule_phase` and
-                        # last_visit_num might not be set yet
-                        yield DetailVariable(name=name, function=form_xpath.first_visit_phase_set)
-                    else:
-                        yield DetailVariable(name=name, function=form_xpath.xpath_phase_set)
-
-            yield DetailVariable(name=SCHEDULE_GLOBAL_NEXT_VISIT_DATE,
-                                 function=u'date(min({}))'.format(','.join(forms_due)))
-            yield DetailVariable(name=SCHEDULE_NEXT_DUE,
-                                 function=ScheduleFormXPath.next_visit_date(last_visit_dates))
-            yield DetailVariable(name='is_late', function=u'${} < today()'.format(SCHEDULE_NEXT_DUE))
-
-            if len(forms_due) != len(set(forms_due)):
-                raise ScheduleError(_("Your app has multiple forms with the same schedule abbreviation"))
-
-    @staticmethod
-    def detail_variables(module, detail, detail_column_infos):
-        return chain(SuiteGenerator._schedule_detail_variables(module, detail, detail_column_infos),)
-
-    def build_case_tile_detail(self, module, detail, detail_type):
-        """
-        Return a Detail node from an apps.app_manager.models.Detail that is
-        configured to use case tiles.
-
-        This method does so by injecting the appropriate strings into a template
-        string.
-        """
-        from corehq.apps.app_manager.detail_screen import get_column_xpath_generator
-
-        template_args = {
-            "detail_id": id_strings.detail(module, detail_type),
-            "title_text_id": id_strings.detail_title_locale(
-                module, detail_type
-            )
-        }
-        # Get field/case property mappings
-
-        cols_by_tile = {col.case_tile_field: col for col in detail.columns}
-        for template_field in ["header", "top_left", "sex", "bottom_left", "date"]:
-            column = cols_by_tile.get(template_field, None)
-            if column is None:
-                raise SuiteError(
-                    'No column was mapped to the "{}" case tile field'.format(
-                        template_field
-                    )
-                )
-            template_args[template_field] = {
-                "prop_name": get_column_xpath_generator(
-                    self.app, module, detail, column
-                ).xpath,
-                "locale_id": id_strings.detail_column_header_locale(
-                    module, detail_type, column,
-                ),
-                # Just using default language for now
-                # The right thing to do would be to reference the app_strings.txt I think
-                "prefix": escape(
-                    column.header.get(self.app.default_language, "")
-                )
-            }
-            if column.format == "enum":
-                template_args[template_field]["enum_keys"] = {}
-                for mapping in column.enum:
-                    template_args[template_field]["enum_keys"][mapping.key] = \
-                        id_strings.detail_column_enum_variable(
-                            module, detail_type, column, mapping.key_as_variable
-                        )
-        # Populate the template
-        detail_as_string = self._case_tile_template_string.format(**template_args)
-        return load_xmlobject_from_string(detail_as_string, xmlclass=Detail)
-
-    @property
-    @memoized
-    def _case_tile_template_string(self):
-        """
-        Return a string suitable for building a case tile detail node
-        through `String.format`.
-        """
-        with open(os.path.join(
-                os.path.dirname(__file__), "case_tile_templates", "tdh.txt"
-        )) as f:
-            return f.read().decode('utf-8')
+        return DetailContributor(self, self.app, self.modules).get_section_contributions()
 
     @staticmethod
     def get_filter_xpath(module, delegation=False):
